Refactor framework to encapsulate node
